+++ src/components/cc-button/cc-button.js
@@ -167,7 +167,7 @@ export class CcButton extends LitElement {
 
     const delay = (this.delay != null && !this.link) ? this.delay : null;
 
-    const waiting = (this.waiting && !this.link);
+    const waiting = (this.waiting);
 
     // simple mode is default when no value or when there are multiple conflicting values
     modes.simple = !modes.primary && !modes.success && !modes.warning && !modes.danger && !this.link;
@@ -418,8 +418,13 @@ export class CcButton extends LitElement {
         }
 
         .outlined progress,
-        .outlined progress::-webkit-progress-bar {
+        .outlined progress::-webkit-progress-bar{
           background-color: var(--btn-color);
+        }
+
+        .cc-link progress,
+        .cc-link progress::-webkit-progress-bar {
+          background-color: var(--cc-color-text-strong);
         }
 
         progress::-webkit-progress-value,
@@ -522,7 +527,11 @@ export class CcButton extends LitElement {
 
         /* button that looks like a cc-link */
         .cc-link {
+          --btn-color: var(--color-text-strong);
           cursor: pointer;
+          min-height: 2em;
+          overflow: hidden;
+          position: relative;
           text-decoration: underline;
         }